Move getTenantId() from Events to Commands.
