+++ server/src/main/java/io/spine/server/tenant/EventOperation.java
@@ -23,7 +23,7 @@ package io.spine.server.tenant;
 import io.spine.annotation.Internal;
 import io.spine.core.Event;
 
-import static io.spine.core.Events.getTenantId;
+import static io.spine.core.Commands.getTenantId;
 
 /**
  * A tenant-aware operation performed in relation to an event.
